util.h: add macro for allocating memory
The macro was created to pair together allocating memory
and checking if the memory allocated is NULL.
All the existing malloc() calls where replaced by the macro
and error handling was added.

Test cases were also added, and the `Makefile.am` file was
modified to ignore some expected warnings only for the tests.

Signed-off-by: Panagiotis Foliadis <pfoliadis@hotmail.com>