cnc: Add --version flag to display version and commit hash
This patch introduces a new `--version` command-line flag to the cnc
program, allowing users to easily retrieve the current version number
and the Git commit hash associated with the build. When invoked with
`--version` or `-V`, the program will display this information,
providing valuable insights into the exact version and build being
used.

More,
  * Update configure.ac to dynamically set the version using `git describe`
  * Add COMMIT_HASH macro in Makefile.am and configure.ac
  * Modify cnc.c to handle the --version flag and display version information
  * Update help message to include the --version flag

Signed-off-by: Charalampos Mitrodimas <charmitro@posteo.net>