db: postgres: Migrate pg_dump to pg_dumpall && pg_restore to psql
By using "pg_dumpall" instead of "pg_dump" we ensure that ALL
resources of the database are being exported in the backup file, which
is now an SQL script instead of `.dump` due to the nature of
"pg_dumpall". This means that previous issues are now resolved,
i.e.
	* Missing privileges
	* Missing roles

We also migated from "pg_restore" to the standard "psql" command,
running it in the format of,
	psql -h <address> -U <user> -p <port> -f <SQL script>

This, with the addition of "pg_dumpall" solves our major issue of
being unable to run "pg_restore" on an already "restored" database.

Signed-off-by: Charalampos Mitrodimas <charmiro@posteo.net>