util: reintroduce macro for allocating memory
On commit `d76e913` a macro was added that was not compliant
with the GCC standards. This macro was replaced on commit `9988d89`
with an inline function, and is now reintroduced and is aligned
with the GCC-standard.
All of the uses of the inline function are replaced with the
CNC_MALLOC macro and also some tests were revised to reflect
the changes of the macro.

Signed-off-by: Panagiotis Foliadis <pfoliadis@hotmail.com>